Fix compile issues in lesson 4
