+++ lesson_4/.vscode/settings.json
@@ -1,0 +1,6 @@
+{
+    "java.debug.settings.onBuildFailureProceed": true,
+    "java.project.sourcePaths": [
+        "src"
+    ]
+}
