--- lesson_4/examples/A_Constructor.java
@@ -1,13 +1,0 @@
-package lesson_4.examples;
-
-public class A_Constructor {
-
-    A_Constructor(){
-        System.out.println("The constructor is called when you instantiate a class");
-    }
-
-    public static void main(String[] args) {
-        new A_Constructor();
-    }
-    
-}
