--- lesson_4/examples/A_MinimalClass.java
@@ -1,5 +1,0 @@
-package lesson_4.examples;
-
-class A_MinimalClass {
-    
-}
