--- lesson_4/examples/B_ClassMain.java
@@ -1,8 +1,0 @@
-package lesson_4.examples;
-
-public class B_ClassMain {
-    public static void main(String[] args) {
-        B_Class a_class = new B_Class();
-        System.out.println(a_class.getColor());
-    }
-}
