--- lesson_4/examples/C_Vehicle.java
@@ -1,8 +1,0 @@
-package lesson_4.examples;
-
-interface C_Vehicle {
-
-    public void honk();
-
-    void honk(int level);
-}
