--- lesson_4/examples/extends_example/A.java
@@ -1,20 +1,0 @@
-package lesson_4.examples.extends_example;
-
-class A {
-    static final double PI = 3.14159;
-
-    static double getPI() {
-        return A.PI;
-    }
-
-    A() {
-        System.out.println("A constructor");
-    }
-
-    String name = "A class";
-
-    protected String getName() {
-        return name;
-    }
-
-}
