--- lesson_4/examples/extends_example/C.java
@@ -1,11 +1,0 @@
-package lesson_4.examples.extends_example;
-
-public class C extends B {
-
-    String name = "C class";
-
-    C() {
-        super();
-        System.out.println("C constructor");
-    }
-}
