--- lesson_4/examples/interface_example/Animal.java
@@ -1,8 +1,0 @@
-package lesson_4.examples.interface_example;
-
-interface Animal {
-
-    void sleep();
-
-    void backFlip();
-}
