--- lesson_4/examples/interface_example/Elephant.java
@@ -1,16 +1,0 @@
-package lesson_4.examples.interface_example;
-
-public class Elephant implements Animal {
-
-    @Override
-    public void sleep() {
-        System.out.println("Elephant sleep");
-
-    }
-
-    @Override
-    public void backFlip() {
-        System.out.println("Elephant died");
-    }
-
-}
