+++ lesson_4/src/examples/A_Constructor.java
@@ -1,0 +1,13 @@
+package examples;
+
+public class A_Constructor {
+
+    A_Constructor(){
+        System.out.println("The constructor is called when you instantiate a class");
+    }
+
+    public static void main(String[] args) {
+        new A_Constructor();
+    }
+    
+}
