+++ lesson_4/src/examples/A_MinimalClass.java
@@ -1,0 +1,5 @@
+package examples;
+
+class A_MinimalClass {
+    
+}
