+++ lesson_4/src/examples/A_Object.java
@@ -1,0 +1,11 @@
+package examples;
+
+public class A_Object {
+
+    public static void main(String[] args) {
+        // A Object is an instance of a class
+        A_MinimalClass minimalClass = new A_MinimalClass();
+        System.out.println(minimalClass.getClass().getName());
+    }
+    
+}
