+++ lesson_4/src/examples/C_Vehicle.java
@@ -1,0 +1,8 @@
+package examples;
+
+interface C_Vehicle {
+
+    public void honk();
+
+    void honk(int level);
+}
