+++ lesson_4/src/examples/extends_example/C.java
@@ -1,0 +1,11 @@
+package examples.extends_example;
+
+public class C extends B {
+
+    String name = "C class";
+
+    C() {
+        super();
+        System.out.println("C constructor");
+    }
+}
