+++ lesson_4/src/examples/extends_example/Main.java
@@ -1,0 +1,17 @@
+package examples.extends_example;
+
+public class Main {
+
+    public static void main(String[] args) {
+        /*
+         * A a = new A(); System.out.println(a.name);
+         * 
+         * B b = new B("some extra text"); System.out.println(b.name);
+         * System.out.println(b.getName());
+         */
+
+        C c = new C();
+        System.out.println(c.getName());
+
+    }
+}
