+++ lesson_4/src/examples/interface_example/Animal.java
@@ -1,0 +1,8 @@
+package examples.interface_example;
+
+interface Animal {
+
+    void sleep();
+
+    void backFlip();
+}
