+++ lesson_4/src/examples/interface_example/Monkey.java
@@ -1,0 +1,17 @@
+package examples.interface_example;
+
+public class Monkey implements Animal {
+
+    @Override
+    public void sleep() {
+        System.out.println("monkey sleeping zzZZZzz");
+
+    }
+
+    @Override
+    public void backFlip() {
+        System.out.println("monkey did a double back flip!!");
+
+    }
+
+}
